Rename favorites -> favoritesRead to abide to existing naming scheme
