+++ src/reducers/FavoritesReducer.js
